requirements.txt
paramiko~=2.12.0
pytest~=7.2.0
pylint
pycodestyle
flake8
pyeapi
jinja2
PyYAML